+++ app/src/main/kotlin/com/numero/itube/di/FragmentModule.kt
@@ -23,5 +23,5 @@ abstract class FragmentModule {
     abstract fun contributeRelativeFavoriteFragment(): RelativeFavoriteFragment
 
     @ContributesAndroidInjector
-    abstract fun contributePlayerSettingsFragment(): PlayerSettingsFragment
+    abstract fun contributePlayerSettingsBottomSheetFragment(): PlayerSettingsBottomSheetFragment
 }
