+++ app/src/main/res/layout/activity_player.xml
@@ -56,37 +56,6 @@
 
     </android.support.constraint.ConstraintLayout>
 
-    <FrameLayout
-        android:id="@+id/bottomSheetLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:elevation="8dp"
-        app:behavior_hideable="true"
-        app:behavior_peekHeight="0dp"
-        app:layout_behavior="@string/bottom_sheet_behavior">
-
-        <android.support.design.card.MaterialCardView
-            android:id="@+id/bottomSheetCardView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginBottom="-16dp"
-            android:layout_marginTop="8dp"
-            android:clickable="true"
-            android:focusable="true"
-            android:focusableInTouchMode="true"
-            app:cardCornerRadius="16dp"
-            app:cardElevation="8dp">
-
-            <FrameLayout
-                android:id="@+id/bottomSheetContainer"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginBottom="16dp" />
-
-        </android.support.design.card.MaterialCardView>
-
-    </FrameLayout>
-
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
         android:layout_width="wrap_content"
